requirements.txt
# torch
torchsummary
einops
opencv-python
numpy
scikit-learn
matplotlib
tqdm
torchvision
timesformer_pytorch